requirements.txt
Flask
pydantic
python-dotenv